+++ CHANGELOG.md
@@ -153,6 +153,8 @@ Adding a new version? You'll need three changes:
 - Allow configuring a GRPCRoute without hostnames and matches that catch all
   requests.
   [#5303](https://github.com/Kong/kubernetes-ingress-controller/pull/5303)
+- Allow the `ws` and `wss` Enterprise protocol values for protocol annotations.
+  [#5399](https://github.com/Kong/kubernetes-ingress-controller/pull/5399)
 
 ### Changed